修复equals方法类型

+++ jmp/mp-service/src/main/java/com/vpu/mp/service/pojo/shop/member/card/LimitNumCardToVo.java
@@ -95,7 +95,8 @@ public class LimitNumCardToVo extends LimitNumCardVo {
 			storeList = storeList.replaceAll("\\s+", "");
 			storeListArray = storeList.split(",");
 			/** 门店类型 */
-			if (MCARD_STP_BAN.equals(storeListArray[0]) || MCARD_STP_ALL.equals(storeListArray[0])) {
+			if (MCARD_STP_BAN.equals(Byte.valueOf(storeListArray[0])) 
+						|| MCARD_STP_ALL.equals(Byte.valueOf(storeListArray[0]))) {
 				storeListType = Byte.valueOf(storeListArray[0]);
 			} else {
 				storeListType = MCARD_STP_PART;
